server side refactoring, validation with Fluent Validation, queries refactoring and some unuseful stuff removed

+++ WebBetApp/WebBetApp/Main/WebBetQuriesImpl.cs
@@ -2,23 +2,32 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 using WebBetApp.Model.Database;
 using WebBetApp.Model.Database.DatabaseModel;
 using WebBetApp.Model.ViewModels;
 using WebBetApp.Helper;
-using WebBetApp.Main.Validation;
-
+using FluentValidation;
+using Microsoft.AspNetCore.Identity;
+using System.Threading.Tasks;
 
 namespace WebBetApp.Main
 {
     public class WebBetQuriesImpl : IWebBetQueries
     {
         private readonly WebBetDbContext _context;
+        private UserManager<ApplicationUser> _userManager;
+        private readonly IValidator<WebTicket> _ticketValidator;
 
-        public WebBetQuriesImpl( WebBetDbContext context)
+        public WebBetQuriesImpl
+            (
+                WebBetDbContext context,
+                UserManager<ApplicationUser> userManager,
+                IValidator<WebTicket> ticketValidator
+            )
         {
             _context = context;
+            _userManager = userManager;
+            _ticketValidator = ticketValidator;
         }
 
         public IEnumerable<WebMatchOffer> GetMatchesGroupedBySport()
@@ -36,11 +45,11 @@ namespace WebBetApp.Main
                 .ToList();
         }
 
-        public IEnumerable<WebTicket> GetAllTickets(ApplicationUser user)
+        public IEnumerable<WebTicket> GetAllTickets(string userId)
         {
             return _context.Tickets
                 .Include(i => i.TicketMatches)
-                .Where( tm => tm.ApplicationUserId == user.Id)
+                .Where( tm => tm.ApplicationUserId == userId)
                 .Select(t =>
                             new WebTicket
                             {
@@ -55,15 +64,25 @@ namespace WebBetApp.Main
                 .ToList();
         }
 
-        public void PostWebTicketToDb(WebTicket webTicket, ApplicationUser user)
+        public List<string> PostWebTicketToDb(WebTicket webTicket, string userId)
         {
-            using (_context)
+            var response = new List<string>();
+
+            webTicket.UserBalance =  GetUserWalletBalance(userId).Amount;
+            webTicket.User = _context.ApplicationUsers.FirstOrDefault(x => x.Id == userId);
+          
+            var result = _ticketValidator.Validate(webTicket);
+
+            if (!result.IsValid)
             {
-                var balance = GetUserWalletBalance(user).Amount;
-
-                WalletValidation.ValidateWalletBalanceGreaterThenStake(webTicket, balance);
-                TicketValidation.ValidateTicket(webTicket, _context);
-
+                foreach (var error in result.Errors)
+                {
+                    response.Add(error.ToString());
+                }
+                return response;
+            }
+            else
+            {
                 try
                 {
                     var ticket = new Ticket
@@ -74,14 +93,14 @@ namespace WebBetApp.Main
                         StakeWithManipulationCosts = webTicket.StakeWithManipulationCosts,
                         TicketMatches = webTicket.TicketMatches.ToList(),
                         TotalMatchesCoefficient = webTicket.TotalMatchesCoefficient,
-                        ApplicationUserId  = user.Id
+                        ApplicationUserId = userId
                     };
 
                     _context.Tickets.Add(ticket);
 
                     foreach (var match in webTicket.TicketMatches)
                     {
-                       _context.TicketMatches.Add(match);
+                        _context.TicketMatches.Add(match);
                     }
 
                     _context.SaveChanges();
@@ -91,7 +110,9 @@ namespace WebBetApp.Main
                         Amount = -webTicket.Stake,
                     };
 
-                    MakeTransaction(withdrawTransaction, user);
+                    MakeTransaction(withdrawTransaction, userId);
+
+                    return response;
                 }
                 catch (Exception ex)
                 {
@@ -100,39 +121,52 @@ namespace WebBetApp.Main
             }
         }
 
-        public void DeleteTicketFromDb(string ticketCode)
+        public Ticket DeleteTicketFromDb(int ticketId)
         {
-            var ticket = _context.Tickets
-                .Include(tm => tm.TicketMatches)
-                .SingleOrDefault(t => t.TicketCode == ticketCode);
-
-            foreach(var ticketMatch in ticket.TicketMatches)
+            try
             {
-                _context.TicketMatches.Remove(ticketMatch);
+                var ticket = _context.Tickets
+                                .Include(tm => tm.TicketMatches)
+                                .SingleOrDefault(t => t.Id == ticketId);
+
+                if (ticket != null)
+                {
+                    foreach (var ticketMatch in ticket.TicketMatches)
+                    {
+                        _context.TicketMatches.Remove(ticketMatch);
+                    }
+
+                    _context.Tickets.Remove(ticket);
+
+                    _context.SaveChanges();
+
+                    return ticket;
+                }
+                else
+                    return null;
+                
             }
-
-            _context.Tickets.Remove(ticket);
-
-            _context.SaveChanges();
+            catch (Exception ex)
+            {
+                throw ex;
+            }
         } 
 
-        public void MakeTransaction(WebWallet webWalletDeposit, ApplicationUser user)
+        public void MakeTransaction(WebWallet webWalletDeposit, string userId)
         {
-            var balance = GetUserWalletBalance(user).Amount;
-            WalletValidation.ValidateBalanceMustNotBeLessThanZero(webWalletDeposit, balance);
-
             try
             {
                var transaction = new Transaction
                {
                   Amount = webWalletDeposit.Amount,
                   TransactionDate = DateTime.Now,
-                  ApplicationUserId = user.Id
+                  ApplicationUserId = userId
                };
 
                _context.Transactions.Add(transaction);
 
                _context.SaveChanges();
+
             }
             catch(Exception ex)
             {
@@ -140,16 +174,47 @@ namespace WebBetApp.Main
             }
             
         }
-        public WebWallet GetUserWalletBalance(ApplicationUser user)
+
+        public WebWallet GetUserWalletBalance(string userId)
         {
             var balanceAmount =  _context.Transactions
-                    .Where(t => t.ApplicationUserId == user.Id)
+                    .Where(t => t.ApplicationUserId == userId)
                     .Sum(a => a.Amount);
 
             return new WebWallet
             {
                 Amount = balanceAmount
             };
+        }
+
+        public UserDetails GetUserDetails(string userId)
+        {
+            return  _context.ApplicationUsers
+                .Where(x => x.Id == userId)
+                .Select(user =>
+                         new UserDetails
+                         {
+                             UserId = userId,
+                             FullName = user.FullName,
+                             UserName = user.UserName,
+                             UserWalletBalance = GetUserWalletBalance(userId).Amount
+                         }
+                       )
+                .SingleOrDefault();
+        }
+
+        public async Task<IdentityResult> CreateUser(UserRegistration userRegistration)
+        {
+            var applicationUser = new ApplicationUser
+            {
+                UserName = userRegistration.UserName,
+                FullName = userRegistration.FullName,
+                Email = userRegistration.Email
+            };
+
+            var result = await _userManager.CreateAsync(applicationUser, userRegistration.Password);
+
+            return result ;
         }
     }
 }
